Fix: Combined loader.py for diapasef and synchropasef
